move drawing to a separate component
